chore: remove unused StandardVariableSpecification
This is used to parse a sequence of Standard Variable records. That is already being handled correctly.
This class serves no purpose since the record count is not always written just before the record sequence, so
each record or PDU that needs to handle a sequence of records does so independently.